Refactor SQL query and logging logic in middleware
The INSERT_LOG_QUERY string in the middleware's quick_search_query has been refactored as a parameterized query to increase security and performance. Additionally, logging logic was updated to use the IDs of data_source_matches instead of the json string of the whole data object, thus improving the clarity and usefulness of logs.