feat(database): refine change log
Add users table to change log.
Add so that INSERTS are also included in the CHANGE log
Add so that if both old_data and new_data columns are equivalent to {}::jsonb, we skip the insertion -- this is unimportant data.
Remove data where both old_data and new_data are equivalent to {}::jsonb